refactor: extract Framework portal to separate route tree
