+++ assignments-portal/src/pages/SQLPortal.vue
@@ -1,0 +1,9 @@
+<template>
+    <div>
+        <h1>SQL Portal</h1>
+    </div>
+</template>
+
+<script setup>
+
+</script>
